test(evm): reuse helper in deploy local coin
